backend/requirements.txt
fastapi[all]
python-multipart
uvicorn[standard]
pyodbc
pydantic
python-decouple
tortoise-orm
aiosqlite